Move: message_handling.go from package voice to discordmessenger
